Safari-611.1.10/LayoutTests/crypto/subtle/ec-import-jwk-key-export-spki-key-expected.txt

Test importing a JWK ECDH public key and export it in SPKI format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS bytesToHexString(publicKey) is expectedSpkiKey
PASS successfullyParsed is true

TEST COMPLETE
